requirements.txt
transformers==4.12.5
cython 
scipy~=1.7.3
opencv-python~=4.6.0.66
pillow~=9.2.0
scikit-image~=0.19.3
timm~=0.5.4
einops~=0.4.1
pandas~=1.3.5
imgaug
h5py~=3.7.0
av
torch~=1.7.1
torchvision~=0.8.0a0
numpy~=1.21.2
tqdm~=4.64.0
setuptools~=63.2.0
MultiScaleDeformableAttention~=1.0
fvcore~=0.1.5
pycocotools~=2.0
clip~=1.0
PyYAML~=6.0
matplotlib~=3.5.2
requests~=2.28.1
ftfy~=6.1.1
regex~=2022.6.2